requirements.txt
allure-pytest==2.11.1
allure-python-commons==2.11.1
Appium-Python-Client==1.1.0
APScheduler==3.9.1.post1
arrow==1.2.3
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
Authlib==1.0.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
black==22.12.0
Brotli==1.1.0
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coloredlogs==15.0.1
colorlog==6.7.0
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
cx-Oracle==8.3.0
DBUtils==3.0.2
ddddocr==1.4.7
ddt==1.6.0
deepdiff==5.8.1
defusedxml==0.7.1
Deprecated==1.2.13
DingtalkChatbot==1.5.3
Django==4.2.5
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
Faker==13.13.0
Flask==2.1.1
Flask-APScheduler==1.12.4
flatbuffers==2.0.7
funppy==0.5.2
gevent==21.12.0
greenlet==1.1.3.post0
grpcio==1.62.1
grpcio-tools==1.62.1
gunicorn==20.1.0
h11==0.14.0
httpie==3.2.2
httpie-jwt-auth==0.4.0
httprunner==4.3.5
humanfriendly==10.0
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.9.5
jsonpath==0.82
loguru==0.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.2.1
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==1.23.4
onnxruntime==1.12.1
opencv-python-headless==4.6.0.66
openpyxl==3.0.10
ordered-set==4.1.0
outcome==1.2.0
packaging==21.3
pandas==1.5.0
paramiko==2.10.5
pathspec==0.12.1
Pillow==9.2.0
platformdirs==4.2.0
pluggy==1.0.0
protobuf==4.25.3
py==1.11.0
pyasn1==0.4.8
PyAutoGUI==0.9.53
PyAutoIt==0.6.5
pycparser==2.21
pydantic==1.8.2
PyExecJS==1.5.1
PyGetWindow==0.0.9
pygments==2.17.2
PyJWT==1.7.1
PyMsgBox==1.0.9
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pytest==7.1.3
pytest-assume==2.4.3
pytest-check==1.0.10
pytest-html==3.1.1
pytest-metadata==2.0.2
pytest-rerunfailures==10.2
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==6.0.1
redis==4.2.2
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
selenium==3.141.0
sentry-sdk==0.14.4
six==1.16.0
sniffio==1.3.0
some-package==0.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
sympy==1.11.1
toml==0.10.2
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.7.1
tzdata==2022.6
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
websocket==0.2.1
websocket-client==1.4.1
Werkzeug==2.2.2
win32-setctime==1.1.0
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.0.3
xlutils==2.0.0
xlwt==1.3.0
xmltodict==0.12.0
zope.event==4.5.0
zope.interface==5.5.2
